commented out unnecessary bits
Commented out lines 379-380 ... i had restricted the "valid commands" to only those that started with "cmd ...".  This was good for stability, but not necessary for all command line applications.  So this change provides more freedom, but on the flip side, requires the user to do more testing for which combinations of parameters actually work.